Move views outside of app
