+++ apps/puzzlify/views/HostAdventureDashboard.tsx
