+++ apps/puzzlify/views/HostStageView.tsx
