+++ apps/puzzlify/views/ParticipantAdventureDashboard.tsx
