+++ apps/puzzlify/views/ParticipantHintView.tsx
