+++ apps/puzzlify/views/ParticipantStageView.tsx
